+++ src/lib/ThemeToggle.svelte
@@ -15,8 +15,11 @@
 
 <DarkMode bind:theme />
 
-<button on:click={toggle} class="cursor-pointer flex items-center py-1 px-0 bg-transparent border-0 font-[var(--font-primary)] color-[var(--text)]">
-	<div class="i-fa6-solid:{theme === "dark" ? "sun" : "moon"}" />
+<button
+	on:click={toggle}
+	class="cursor-pointer flex items-center py-1 px-0 bg-transparent border-0 font-[var(--font-primary)] color-[var(--text)]"
+>
+	<div class="i-fa6-solid:{theme === 'dark' ? 'sun' : 'moon'}" />
 	<span class="ml-2">Toggle theme</span>
 </button>
