+++ src/lib/fetchState.ts
@@ -1,0 +1,7 @@
+let state: any = [];
+export default async function fetchState(module: "announcements" | "form") {
+	const url = `https://segfautils.projectsegfau.lt/api/set/${module}`;
+	const response = await fetch(url);
+	state = await response.json();
+	return state;
+}
