+++ src/routes/legal/tos/+page.md
@@ -14,5 +14,6 @@ description: We're not anarchists, so we have some rules.
 5. While we do try to keep your data safe, you have to acknowledge that **we are not responsible** if anything **unintentional** happens (such as **data loss**, **inability to extract your data** due to the server being down.). It is also **your responsibility** to **keep a backup of your data** if it matters to you.
 
 ### Failure to comply will either end up **disabling** your account, and/or if it is illegal, **reporting** to the police.
-* Be aware that "**reporting** to the police" only happens if the individual is using our services to **post, promote, or encourage** actions that are **morally unnacceptable**. (Such as: pedophilia, murder, or CSAM.)
-* **No warnings will be issued if we find this out.**
+
+-   Be aware that "**reporting** to the police" only happens if the individual is using our services to **post, promote, or encourage** actions that are **morally unnacceptable**. (Such as: pedophilia, murder, or CSAM.)
+-   **No warnings will be issued if we find this out.**
