+++ src/routes/timeline/+page.svelte
@@ -232,11 +232,15 @@
 			<TimelineConnector />
 		</TimelineSeparator>
 		<TimelineContent>
-			<h3>
-				Resignation of members and the start of a new era.
-			</h3>
+			<h3>Resignation of members and the start of a new era.</h3>
 			<p>
-				Devnol, Odyssey and akisblack resigned from the team leading to a serious loss of manpower. To combat this, we decided to shut down some of our unused instances as well as the ones that were painful to deal with. You can read more over at <a href="https://blog.projectsegfau.lt/resignation-of-3-members-of-projectsegfault/">this blog post</a>.
+				Devnol, Odyssey and akisblack resigned from the team leading to
+				a serious loss of manpower. To combat this, we decided to shut
+				down some of our unused instances as well as the ones that were
+				painful to deal with. You can read more over at <a
+					href="https://blog.projectsegfau.lt/resignation-of-3-members-of-projectsegfault/"
+					>this blog post</a
+				>.
 			</p>
 		</TimelineContent>
 	</TimelineItem>
